btcaccumulator: Init
btcaccumulator is a package that wraps the utreexo accmulators to be
more convenient to use.  UData is all the data that is needed to prove a
bitcoin block with only the roots of the utreexo accumluator. LeafData
represents all the data that is gonna be hashed and committed into the
accumulator.