wire: Change max block payload and protocol version
Max block payload has been raised to 12MB to account for the utreexo
proof size.  Protocol version has been set to 170013 (just a 1 stuck at
the front of the current btcd supported protocol version so other nodes
don't mistaken this for something else).

All changes here are experimental and is subject to change.