blockchain/indexers, config: Refactor utreexoproof related code
The new refactor adds missing comments, gets rid of the todo global
variable used to store the dataDir, and changes the
NewUtreexoProofIndex() function to initialize the UtreexoState, getting
rid of the need for the caller to initalize it.

New flag checks are included so that the user can't turn on conflicting
utreexo related flags.