wire, blockchain: Use UData compact serialization for MsgBlock
The UData storage and the network serialization for MsgBlock is
changed to the compact serialization.  This results in massive reduction
in storage and bandwidth usage as blockhash and outpoint took up 68
additional bytes per txo.

Since the UData needs to be made full for generating the hash
commitments, a new function generateCommitments() is added that takes in
compact UData along with other information to generate the hashes and
make the UData full.