blockchain, mempool, wire: Use compact UData serialization for MsgTx
Compact UData serialization is now used for propagating transactions
between Utreexo enabled nodes.  This has massive savings in bandwidth
used as blockhash and outpoint took up 68 additional bytes per input.

VerifyUData() in blockchain is modified to account for the compact
UData serialization (for accumulator verification).

filterMempoolTxs() is added in mempool to filter out validating UData
for inputs that are not yet included in a block.