wire: New leaf compact serialization format
With leaf compact serialization, the outpoint is left out.  Because of
this, the reader is responsible for filling in that information.
However, with tx messages, inputs may be unconfirmed, resulting in no
leaf data for such inputs.

This caused a problem where you might have 3 inputs but only 2 leaf
datas.  Since the compact leaf data leaves out the outpoint, you don't
know which inputs correspond to which leaf data.

The new serialization method fixes this by forcing the sender to always
send the equal amount of leaf datas for each of the inputs.  If an input
is unconfirmed, you simply send a 0x01, and the reader recognizes this
as leaf data corresponding to an unconfirmed input.

Note that only tx messages are effected and block messages stay the same.

Lastly, the unneccesary block height information from UData is removed.