wire/batchproof: Add batchproof.go
BatchProof serialization was defined in github.com/mit-dci/utreexo.
However, this BatchProof serialization did not use variable integers,
therefore making UData messages sent over the wire to be bigger.

Redefining the BatchProof serialization in github.com/mit-dci/utreexo is
tricky since there is no variable integer support in that repo.  We
couldn't import the wire package into the utreexo repo since that caused
a dependency cycle.

The solution I came to is to just define a separate function for
serialization in this repo and I think this is a good enough solution.