rpcserver: Add gettxtotals command
gettxtotals returns the total bytes sent and received for all peers for
just the tx messages.  It also calculates the size of the utreexo proof
that would have been needed to prove that tx.

The main functinality of this rpc call is to calculate how much the
utreexo proofs would increase the network traffic needed to run a
utreexo node.