btcutil: Make btcutil part of the btcd repo
btcutil was maintained as a separate repo for btcsuite.  However, there
really is no reason to do this as the package can be imported separately
while it is in the same repo.  Furthermore, btcutil itself depends on
btcd packages so it can't be a standalone package anyways.

btcutil package is copied from commit e2ba6805a89060647fa22ec61b799acff6a4f6fd