blockchain/indexers: Add test for utreexo indexers and fix bugs
A test is added that tests for reorgs, make sure the indexers are
agreeing on the data stored, and that a compact utreexo node is able to
consume the data sent by the utreexo indexers.

Bugs that effected reorgs and cases where a compact utreexo node
couldn't consume the data created by the indexers are fixed.