wire/udata: Explicitly serialize the leaf data count.
There exists an edge case where the leaf data count will not match the
batchproof target count.  When the leaf being deleted is a root, there
will not be any proof.  However, we still need to provide the leaf data
for that root for utxo validation (validation of the signature, amount,
etc).

The new serialization for both the compact and normal udata will include
a varint serialized count of the leaf data before the actual leaf data.

The check for targets and leaf data count equality in the function
ProofSanity() is now removed as there is no guarantee that they will be
equal.