blockchain/utreexoviewpoint: Refactor Modify() function
Functionality that was included in Modify() is now refactored into 3
separate functions: Modify(), IngestProof(), and ExtractAccumulatorAddDels().

Modify now only modifies the accumulator without any other checks.
IngestProof validates and ingests the proof so that the caller can call
Modify().
ExtractAccumulatorAddDels now does the proof sanity checks and extracts
the additions and deletions from the block that was passed in.