wire/udata: Fix bugs, add remember idx, new serialize function
Size serialization functions were incorrect and those are corrected now.
New test add sanity checking for the hardcoded sizes.

Remember indexes are added to UData.  These serve as the indexes of the
UTXOs that should be remembered in a given block.  If the UTXO index
matches up with the index provided here, the leaf should be remembered
by the CSNs.

UData serialization without the accumulator proof is added.  This allows
the caller to leave out the accumulator proof without also leaving out
the targets (which are a list of uint64s that represent the positions of
the leaves to be deleted in the accumulator).