blockchain/indexers: Support multi-block proof indexing
FlatUtreexoProofIndex now supports multi-block proof indexing.  Bug
fixes have been applied from the previous multi-block proof
implementation and tests have been added to ensure CSNs are able to
verify the proof and perform the initial block download with the
multi-block proofs.

Differing multi-block proof intervals cannot be stored together at the
same time at the moment.  In the future you could hardcode some height
where the proof interval changes.  At the moment, more exploring needs
to be done to see which interval is optimal.

UtreexoProofStats keeps track of relevant data for accumluator proof
sizes.  These are currently hardcoded to be printed every 1000 blocks.
The serialization of the stats are straightforward and are stored in
flatfiles along with the block height.  One could just keep this file
and parse it to get the stats.