blockchain: adding bestheader variable to the blockindex
best header tracks the highest work block node in the index that is not known to be invalid. This is specifically useful when we don't have the block data. However, since block nodes are only added to the index for block headers that pass all sanity and positional checks, which include checking proof of work, it does represent the tip of the header chain with the highest known work that has a reasonably high chance of becoming the best chain tip and is useful for things such as reporting progress and discovering the most suitable blocks to download.