blockchain: Add utreexo data to Block
The utreexo update data and the utreexo add hashes are added to the
Block data structure, allowing for easy access to those fields for other
packages receiving a block.

This specifically allows for wallet support for utreexo compact state
nodes as the wallet needs to keep track of its own proofs. The update
data and the add hashes are need to update and reorg the wallet proofs.