wallet, wire: Add watchonly wallet package
Watchonly wallet supports registering addresses and xpubs which will
then be kept track of by the watch only wallet manager. It will keep
track of the relevant utxos and the utreexo proof for all the relevant
utxos for the wallet.

Currently only watching addresses are implemented. Registering xpubs are
implemented but the watch only wallet manager does not keep track of
them yet.

In the wire package, msgblock serialization/deserialization is change so
that it encodes utreexo blocks with the utreexo data and deserialization
tries to decode everything as a utreexo block. This is done so that
compact state nodes will both store and fetch blocks with the utreexo
proofs. The behavior of none utreexo nodes are not affected by this
change.