database/ffldb: Add getCurrentFileNum function
This change is part of the effort to add pruning support to utreexod.

getCurrentFileNum returns the in-memory saved file number of the block
store it's pointing to. Since each spend journal entry refers to a block,
we can remove the spend journal entry when that block is removed. To do
this, the spend journal files should match up with the block files
(ie 000000001-undo.fdb should contains the spend journals for blocks that
are present in 000000001.fdb). We'll need to fetch the current file
number to save the spend journals in such a way.