database/ffldb: Decrease file size to 128MiB from 512MiB
This change is part of the effort to add pruning support to utreexod.

While the original size in btcd was 512MiB, the block size is lowered
in utreexod to 128MiB. Since block files are allocated and deleted per
file, we can lower the minimum requirement by lowering the file size.

Preivously the minimum requirement would be at least 512MiB + spend
journal files. Now the requirement is lessened to 128MiB + spend journal
files.