database/ffldb: Add block location arg to writeBlock
This change is part of the effort to add pruning support to utreexod.

It's beneficial to have the spend journals stored in an undo file to
match up to the blk file (ie 000000001-undo.fdb contains the spend
journals for the blocks that are present in 000000001.fdb). This change
allows us to delete the matching spend journal file when deleting the
blk file.

To support this, writeBlock is changed so that the caller can ask it to
store the data into the file location it requested.