database/ffldb: Add block height bucket
This change is part of the effort to add pruning support to utreexod.

The block height bucket stores the lowest block height and the highest
block height for a given file number. The file number isn't
distinguished between blk and undo files as there will be 1 undo file
per blk file.