database/ffldb: Add fileSizeFunc and fileStartEndNumFunc to blockStore
This change is part of the effort to add pruning support to utreexod.

These two functions are added as fields so that whitebox_test.go can
also implement them.

fileSizeFunc is added so that we can get the write cursor offset when
writing into an arbitrary file.  This is needed as the spend journal
writes will not always be in order.  We want to support pruning while
also supporting parallel block downloads so the blocks may be stored out
of order.  But for easy pruning, it's guaranteed that the spend journal
and the block will be stored in the same file number.  So when there are
spend journal writes into a different file than the one we're currently
on, we'll need to recover the write cursor's offset and fileSizeFunc
provides that functionality.

fileStartEndNumFunc is added so that we can figure out the first and
the last file number on the disk.  Before pruning, it's guaranteed we'd
start from 0.  But with pruning, we don't have that guarantee anymore so
we need this functionality.