database/ffldb: write height data for block files
This change is part of the effort to add pruning support to utreexod.

When writing the block files to the database, we also store the first
and the last heights of the block into the database.  This allows us to
keep certain block files when pruning and allows us to tell the caller
what the earliest block that's stored on the database is.