database/ffldb: Store spend journal data in the same file number as the block
This change is part of the effort to add pruning support to utreexod.

Block and spend journal data were not guaranteed to be in the same file
number on the database.  This change ensures that they are.  When
storing the spend journal data, we'll first check where the block data
is stored and store it in the same file number as the block.  This
allows us to delete a spend journal file that has the same number as the
block file during pruning.