blockchain: Set last flush hash during a cache flush
This change is part of the effort to add pruning support to utreexod.

When pruning is on, the last flush block for the utxo cache determines
if the utxo cache should be flushed.  This is so that on an unexpected
shutdown, the state of the utxo set on the disk is after the earliest
block we have on this.  This ensures that we're able to repair the
database on unclean shutdowns.