blockchain: Add pruning support
This change is part of the effort to add pruning support to utreexod.

A field for pruning is added and the BlockChain struct is now able to be
configured with pruning.  Prune is called on every block connect and
the prune target field is passed to PruneBlocks func.  Lastest 288
blocks are guaranteed to exist on the database so we'll be able to
support the NODE_NETWORK_LIMITED service.

Utxo cache is flushed when the earliest block height that the database
has is earlier than the last utxo cache flush so that the database is
able to recover on unclean shutdowns.