blockchain: Add ExtractMerkleBranch and HashMerkleRoot
Two merkle tree functions are added. ExtractMerkleBranch let's you
extract the merkle proof for a single hash in a merkle tree.
HashMerkleRoot provides a function for calculating the merkle root given
the hash, position, and the merkle branch.

This commit is part of the effort to add electrum server support to
utreexod. Electrum protocol requires serving merkle proofs for a given
tx and the extract provides that functionality. HashMerkleRoot is added
to test that the returned result from ExtractMerkleBranch is correct.