wallet: Make the wallet compatible with electrum
The wallet is modified to support electrum. It supports utreexo and
electrum protocol v1.4.1 as it stores has all the data needed for the
two protocols.

The current method of storage is with json files. The original intention
was that the wallet state would be simple and small enough for the user
to edit and import wallets from outside with json.  However, the
electrum support adds a lot of extra data and the json file is a bit too
big for a user to easily edit it. The method of storing the wallet state
may change in the future.