main: Add flags to allow extended pubkey registration
--registerextendedpubkeystowatchonlywallet is added so that extended
pubkeys can be added to the watch only wallet. This is useful for
wallets that allow for the exporting of the extended pubkey.

--registerextendedpubkeyswithaddresstypetowatchonlywallet is added along
with the other flag as some wallets don't follow the bip/slip specified
hd version bytes. This allows the user to specify which path the wallet
should be generating.