electrum: Add electrum server
The added electrum server is stateless and all the information it serves
is fetched either from the watchonly wallet or the node's own
mempool/blockchain data.

The server supports electrum protocol version v1.4.1 and is tested to be
fully compatible with the electrum desktop wallet (electrum.org) and the
nunchuk wallet (nunchuk.io). Other wallets should work but were not
tested.