mempool, main: cache tx proofs for CSNs
For utreexo CSN to CSN proof serving, each node needs to keep track of
the proof for the given mempool tx. The change made here introduces a
pool of leaf datas that are stored in the mempool and when adding
transactions to the pool, the accumulator proof for that is cached in
the accumulator as well.

This code is mostly for the p2p to eventually support receiving and
serving partial proofs for utreexo peers.