peer, main: allow for queuing multiple invvects
For utreexo tx inv messages, the invvects will be related to each other
as the invvect for the tx will be matched with the position invs. It'll
look like so:

[InvTypeUtreexoTx][TXID1]
[InvTypeProofHash][packed positions1.1] (4 positions)
[InvTypeProofHash][packed positions1.2] (1 position)

At the moment the invvects can only be queued one at a time which may
result in the InvTypeProofHash to be separated from the
InvTypeUtreexoTx. Allowing for queuing multiple invvects eliminates this
problem.