peer: support queuing utreexo proof hash invs
Utreexo proof hash invs are always paired with the tx invs and the code
change is made to handle this. Each utreexo tx is also queued
immediately instead of getting pushed back into the queue to avoid
sending multiple tx invs to utreexo peers. This is done on purpose so
that there isn't any confusion on which utreexo proof hash inv is for
which tx. Multiple txs will be supported via a different wire message in
the future.