mempool, main, electrum, netsync: bugfix for utreexo txs on rbf
For rbf, the utreexo proof that was needed was nothing since the proof
is already cached. However, the removeTransaction that was being called
before the replacement was uncaching the already cached proof, leading
to ingestion failure. To handle this, there's a flag introduced to
removeTransaction to opt out of removing the utreexo proof for a given
tx.