blockchain: hardcode unspendable leafhashes (bip30 violations)
Since utreexo leaf hashes have the block hash committed into the
preimage, there's no overwriting of the utxos of the historical bip30
violations. But since non-utreexo nodes consider them unspendable, we
must also state them as unspendable to avoid consensus
incompatibilities.

This commit just hardcodes the leaf hashes so they can be used. The
added tests are just sanity checking the hardcoded hashes. It also makes
it easier to verify that the hashes are indeed correct.