indexers, main: change pruning behavior with flatutreexoproofindex
The flatutreexoproofindex will now generate undo files only when the
node is pruned. It will also not keep the proof data when it is pruned.

The behavior of the bridge nodes will be to only serve the historical
proofs if it isn't a pruned node. This better reflects the bridge node's
role as a proof-generator for new blocks/transactions and not a node
that just serves proofs for historical blocks.

Since the proofs can also provide the data for undo, there's no need to
keep a separate undo data when the node is archival. So to save on data
storage, the flatutreexoproofindex will only create proofs when the node
is pruned and/or if it's configured to do multi-block proofs.