indexers, main: changed pruning behavior for utreexoproofindex
Utreexoproofindex now deletes historical proofs and starts keeping undo
data for pruned nodes.

This behavior change better reflects the roles of bridge nodes as
generators of proof for new block/txs and not a role of serving
historical blocks/proofs.

Since proofs can be used to provide the necessary undo data, the undo
data isn't created when the utreexoproofindex is in archival mode.