main: add support for rebroadcastunconfirmedbdktxs command
It rebroadcasts the unconfirmed txs that are stored on the bdk side to
peers connected to the node. A node could be shut down after it has
passed off the tx to bdk but before it has sent the txs off to other
peers. Having this rpc makes it so that the user can manually trigger
rebroadcasts.