netsync: support for building the entire headers chain first
Since we need the headers for validating timelock based txs (csv, cltv),
we need to build the entire headers first in order to then initialize
and sync from the assumed utreexo state. The code change here will check
to see if the assumed utreexo point is active and then choose to build
headers and then start syncing from the assumed utreexo point.