addrmgr: fix testing bug that happens here and there
There's a bug that happens occasionally that I'm pretty sure is
happening because these tests are running in parallel. Give the tmp
directories different paths to fix the bug.

The error is always:
`addrmanager_internal_test.go:165: expected to find 5 adddresses, found 4`

If anything it's ioutil.TmpDir is deprecated so this is a good change.