blockchain: use GetReconstructScriptType instead of GetScriptClass
Since GetScriptClass doesn't look at the sigscript or the witness when
determining the type, it'll still consider p2pkh scripts that have
OP_DROP or any other uncommon sigscripts/witnesses as p2pkh. This can't
be considered safe since we don't really know what's possible with these
scripts so we prefer GetReconstructScriptType, which does take a look at
the sigscript and the witness and considers any sigscript that's not
structured like [signature pubkey] as not reconstructable.