main: don't connect to non-utreexo peers as a csn
A csn would previously connect to any node that signaled NODE_NETWORK.
However, since a csn can't sync or verify txs from such nodes, it's not
feasible to have them as peers to receive txs or blocks from.

However, it's possible to give txs/blocks to these peers. But we'll get
to that at a later time.