wallet: fix how leafdatas are fetched
When the mempool didn't have caching built in, it used to have the full
udata attached to the tx. But since this is not the case anymore,
watchonly wallet panics as it tries to access something that no longer
exists.

We change the code to fetch the leafdatas from the mempool directly as
it can now be fetched from there.