wire, main, indexers: remove unconfirmed marker
The unconfirmed marker created a discrepancy between the proof sent out
for blocks and transactions. It was used for marking if a referenced tx
that was being spent was unconfirmed or not but it's removed in favor of
the node checking locally for the referenced tx. If the node is unable
to find all the referenced txs, the tx is put into the orphan pool.