indexers: write consistency state on flushes
When the utreexo state is flushed to disk, the best block hash and the
numleaves are also written. This is all done in one leveldb transaction
and so it's guaranteed to be atomic.

The best block hash that's written allows for callers to check if the
utreexo state is consistent or not with the blockindex.