indexers: don't init utreexo state when creating a new index
In order to support checking if the utreexo state is consistent, we need
the best block hash, which can only be attained when the chain is
initialized. The current initialization for the utreexo state was being
done when the index was created. Moving the initialization to when the
chain is being initialized allows us to pass in the best block hash.