indexers: require tip hash and height for index Init
This change requires the indexers to take in the tip hash and height as
an argument to Init(). With this, utreexo indexes are able to see what
tip they're synced to. Since the accumulator isn't written to disk every
block, this is useful information to check which block each utreexo
index should sync to.