indexers: flush utreexo state on block disconnects
Always flush the utreexo state on block disconnects. This allows the
utreexo state to always stay at a recoverable state. Even if the ffldb
fails to write, we can always re-connect blocks and reorganize. However,
if the utreexo state isn't flushed, we can't disconnect on restart as
the data could have been deleted from ffldb.