blockchain, indexers: add flush to indexer interface
Flush is added to the indexer interface so that we can call the flush
from when connecting the block in the blockchain package. This allows
the utreexo indexes to also flush the main database. They couldn't when
being called from the ConnectBlock on the indexers as that acquires the
database tx lock.