blockchain, indexers, netsync: add flush for Indexers
FlushIndexes method is added to blockchain so that external callers can
trigger flushes on the indexes. It's useful for periodic flushes were
the node is already caught up to the tip but should be flushed to keep
the node from being too far behind if there were to be an unexpected
shutdown.