indexers: flush utreexo state on prune
For the utreexo state to be recoverable on unexpected crashes, there
must be blocks available for it to reindex on crashes. If there aren't,
the utreexo state is irrecoverable. To prevent this from happening, we
check what the last stored block on the disk is after a prune and flush
the utreexo state if the last flush happened before the last kept block.