utreexobackends: add CachedPosition
When mapping to a uint64, it wasn't possible to mark if the positions
were fresh or not. If a position could be marked fresh, then we can
delete it completely from the memory without it having to touch the
disk.

This change comes from observing that there's a lot of slowdowns coming
from flushing.