blockchain: change nodesbackend deletion
For NodesBackend, the cached entry is not removed from the cache even
if it's being deleted as subsiquent calls would be made to fetch the
key. Caching it as removed saves on disk reads and on flushes, the leaf
is not attempted to be flushed if it's marked as fresh.