netsync, peer: never add invs with the utreexo flag
Since we check for the entire invvect to see if a peer has a particular
tx or block, we can mistake that the peer doesn't have a tx or a block
when they do if the invtype has the utreexo flag.

Since having the utreexo flag doesn't change the underlying tx or the
block, it's better to not distinguish between the two when tracking if
the peer has a certain block/tx.