blockchain: don't assign header as the best tip on maybeAcceptBlockHeader
maybeAcceptBlockHeader verified the passed in block header and set it as
the best tip on it passing verification. However, since the best tip is
used for marking a block that's passed all verification and headers
verification isn't complete verification, we remove the code for adding
the blockNode as the best tip.

This allows us to use the ProcessBlockHeader code for verification of a
header without accepting it as well.