blockchain: change HaveBlock behavior
Since ProcessBlockHeader will add the block header as a blockNode to the
blockIndex, the HaveBlock function will incorrectly state that the block
exists when only the header exists.

Because of this, we check that the node exists and also check the node's
status to see if the data is stored.